restructure folder for gradr
