+++ src/estimator.js
